projects/agentic-ai/game-answer-agent/requirements.txt
chromadb>=1.0.4
openai>=1.73.0
pydantic>=2.11.3
python-dotenv>=1.1.0
tavily-python>=0.5.4
pdfplumber
json-repair